% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.boss.R
\name{cv.boss}
\alias{cv.boss}
\title{Cross validation for BOSS.}
\usage{
cv.boss(x, y, n.folds=10, n.rep=1, ...)
}
\arguments{
\item{x}{A matrix of predictors, see \code{boss}.}

\item{y}{A vector of response variable, see \code{boss}.}

\item{n.folds}{The number of cross validation folds.}

\item{n.rep}{The number of replications of cross validation.}

\item{...}{Arguments to \code{boss}.}
}
\value{
\itemize{
  \item boss: An object \code{boss} that fits on the full dataset.
  \item n.folds: The number of cross validation folds.
  \item cvm.fs: Mean OOS deviance for each candidate given by FS.
  \item cvm.boss: Mean OSS deviance for each candidate given by BOSS.
  \item i.min.fs: The index of minimum cvm.fs.
  \item i.min.boss: The index of minimum cvm.boss.
}
}
\description{
Cross validation for BOSS and FS.
}
\details{
This function fits BOSS and FS (\code{boss}) on the full dataset, and performs \code{n.folds}
  cross validation. The cross validation process can be repeated \code{n.rep} times to evaluate the
  out-of-sample (OOS) performance for the candidate subsets given by both methods.
}
\examples{
# Generate a trivial dataset, X has mean 0 and norm 1, y has mean 0
set.seed(11)
n = 20
p = 5
x = matrix(rnorm(n*p), nrow=n, ncol=p)
x = scale(x, center = colMeans(x))
x = scale(x, scale = sqrt(colSums(x^2)))
beta = c(1, 1, 0, 0, 0)
y = x\%*\%beta + scale(rnorm(20, sd=0.01), center = TRUE, scale = FALSE)


# perform 10-fold CV without replication
boss_cv_result = cv.boss(x, y, n.folds = 2)
boss_result = boss_cv_result$boss
# select the best candidate based on minimum CV OSS score
beta_boss_cv = boss_result$beta_boss[,boss_cv_result$i.min.boss]
# we can compare with the subset selected via IC
# AICc
beta_boss_aicc = boss_result$beta_boss[,which.min(boss_result$IC_boss$aicc)]
# Mallow's Cp
beta_boss_cp = boss_result$beta_boss[,which.min(boss_result$IC_boss$cp)]

}
\references{
Tian, Hurvich and Simonoff (2018), On the use of information criterion
  in least squares based subset selection problems. (Link to be added)
}
\author{
Sen Tian
}
