% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boss.R
\name{predict.boss}
\alias{predict.boss}
\title{Prediction given new data entries.}
\usage{
\method{predict}{boss}(object, newx, ...)
}
\arguments{
\item{object}{The boss object, returned from calling 'boss' function.}

\item{newx}{A new data entry or several entries. It can be a vector, or a matrix with
\code{nrow(newx)} being the number of new entries and \code{ncol(newx)=p} being the
number of predictors. The function takes care of the intercept, NO need to add \code{1}
to \code{newx}.}

\item{...}{Extra arguments to be plugged into \code{coef}, such as \code{select.boss},
see the description of \code{coef} for more details.}
}
\value{
\itemize{
  \item fs: The prediction(s) for FS.
  \item boss: The prediction(s) for BOSS.
}
}
\description{
This function returns the prediction(s) given new observation(s), for FS and BOSS,
where the optimal coefficient vector is chosen via certain selection rule.
}
\details{
The function basically calculates \eqn{x * coef}, where \code{coef}
is a coefficient vector chosen by a selection rule. See more details about the default
and available choices of the selection rule in the description of \code{coef.boss}.
}
\examples{
#See the example in the section of \\code{boss}. Or type ?boss in R.

}
