% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.boss.R
\name{predict.cv.boss}
\alias{predict.cv.boss}
\title{Prediction given new data entries.}
\usage{
\method{predict}{cv.boss}(object, newx, ...)
}
\arguments{
\item{object}{The cv.boss object, returned from calling 'cv.boss' function.}

\item{newx}{A new data entry or several entries. It can be a vector, or a matrix with
\code{nrow(newx)} being the number of new entries and \code{ncol(newx)=p} being the
number of predictors. The function takes care of the intercept, NO need to add \code{1}
to \code{newx}.}

\item{...}{Extra arguments (unused for now).}
}
\value{
\itemize{
  \item fs: The prediction for FS.
  \item boss: The prediction for BOSS.
}
}
\description{
This function returns the prediction(s) given new observation(s), for FS and BOSS,
where the optimal coefficient vector is chosen via cross-validation.
}
\examples{
# See the example in the section of \\code{cv.boss}. Or type ?cv.boss in R.

}
