% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.boss.R
\name{coef.cv.boss}
\alias{coef.cv.boss}
\title{Select coefficient vector based on cross validation (CV).}
\usage{
\method{coef}{cv.boss}(object, ...)
}
\arguments{
\item{object}{The cv.boss object, returned from calling 'cv.boss' function.}

\item{...}{Extra arguments (unused for now).}
}
\value{
\itemize{
  \item fs: The chosen coefficient vector for FS.
  \item boss: The chosen coefficient vector for FS.
}
}
\description{
This function returns coefficient vector that minimizes out-of-sample (OOS) cross
validation score.
}
\examples{
# See the example in the section of \\code{cv.boss}. Or type ?cv.boss in R.

}
